safari-534.53-branch/LayoutTests/fast/dom/Geolocation/permission-denied-expected.txt

Tests Geolocation when permission is denied, using the mock service.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.code is error.PERMISSION_DENIED
PASS error.message is "User denied Geolocation"
PASS successfullyParsed is true

TEST COMPLETE
